@@ -70,24 +71,24 @@
 		"io-ts": "^2.2.18",
 		"katex": "^0.16.2",
 		"mdast": "^3.0.0",
-		"moment": "^2.24.0",
-		"monocle-ts": "^2.1.0",
-		"newtype-ts": "^0.3.4",
-		"query-string": "^6.4.0",
-		"react": "^16.8.4",
-		"react-color": "^2.17.0",
-		"react-dom": "^16.8.4",
-		"react-google-recaptcha": "^1.0.5",
+		"moment": "^2.29.4",
+		"monocle-ts": "^2.3.13",
+		"newtype-ts": "^0.3.5",
+		"query-string": "^7.1.1",
+		"react": "^18.2.0",
+		"react-color": "^2.19.3",
+		"react-dom": "^18.2.0",
+		"react-google-recaptcha": "^2.1.0",
 		"react-helmet-async": "^1.3.0",
-		"react-modal": "^3.8.1",
-		"react-redux": "^7.1.3",
-		"react-rnd": "^9.1.2",
-		"react-router": "^5.0.1",
-		"react-router-dom": "^5.0.1",
-		"react-twitter-embed": "^2.0.7",
-		"react-use": "^14.2.0",
-		"redux": "^4.0.4",
-		"redux-observable": "^1.2.0",
+		"react-modal": "^3.15.1",
+		"react-redux": "^8.0.2",
+		"react-rnd": "^10.3.7",
+		"react-router": "^5.3.3",
+		"react-router-dom": "^5.3.3",
+		"react-twitter-embed": "^4.0.4",
+		"react-use": "^17.4.0",
+		"redux": "^4.2.0",
+		"redux-observable": "^2.0.0",
 		"rehype-highlight": "^5.0.2",
 		"rehype-katex": "^6.0.2",
 		"rehype-react": "^7.1.1",
