@@ -1,4 +1,4 @@
-import { Icon, IconButton } from "@material-ui/core";
+import { Icon, IconButton } from "@mui/material";
 import * as React from "react";
 import { useToggle } from "react-use";
 import * as G from "../generated/graphql";
