@@ -1,5 +1,5 @@
 import * as routes from "@anontown-frontend/routes";
-import { Icon, IconButton, Paper } from "@material-ui/core";
+import { Icon, IconButton, Paper } from "@mui/material";
 import * as React from "react";
 import { Link } from "react-router-dom";
 import * as G from "../generated/graphql";
