@@ -9,11 +9,10 @@ import { configureStore } from "./domains";
 import { gqlClient } from "./effects";
 import "./main.scss";
 import { HelmetProvider } from "react-helmet-async";
-import { createTheme, MuiThemeProvider } from "@material-ui/core/styles";
+import CssBaseline from "@mui/material/CssBaseline";
 
 Modal.setAppElement("#root");
 
-const theme = createTheme({});
 const store = configureStore();
 
 ReactDOM.render(
@@ -21,9 +20,8 @@ ReactDOM.render(
     <Provider store={store}>
       <HelmetProvider>
         <BrowserRouter>
-          <MuiThemeProvider theme={theme}>
-            <App />
-          </MuiThemeProvider>
+          <CssBaseline />
+          <App />
         </BrowserRouter>
       </HelmetProvider>
     </Provider>
