@@ -1,4 +1,4 @@
-import { Paper, Button, TextField } from "@material-ui/core";
+import { Paper, Button, TextField } from "@mui/material";
 import * as React from "react";
 import { Helmet } from "react-helmet-async";
 import { Errors, Page, Snack } from "../components";
