@@ -1,4 +1,4 @@
-import { Paper } from "@material-ui/core";
+import { Paper } from "@mui/material";
 import * as React from "react";
 import { RouteComponentProps, withRouter } from "react-router-dom";
 import { Page, Snack, TopicData } from "../components";
