@@ -1,11 +1,12 @@
-import useRouter from "use-react-router";
 import * as React from "react";
+import { useHistory, useLocation } from "react-router";
 
 export function useOnChnageUrlSearch<A>(
   parse: (query: string) => A,
   onChange: (x: A) => void
 ): A {
-  const { history, location } = useRouter();
+  const history = useHistory();
+  const location = useLocation();
 
   React.useEffect(() => {
     return history.listen((e, action) => {
